@@ -1,12 +1,12 @@
 name: simple_result
-version: 0.1.9
+version: 1.0.0
 homepage: https://github.com/spielgestalt/simple_result.git
 description: A simple posibility to return a Result with eiter success or failure.
 environment:
-  sdk: ">=2.4.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
+  equatable: ^2.0.0
   meta: ^1.1.8
-  equatable: ^1.2.2
 dev_dependencies:
   lint: ^1.1.1
   test: ^1.12.0  
