@@ -1,17 +1,17 @@
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 
 typedef FlatMapFunction<A, B> = int Function(double a);
 
 /// A simple Result wrapper.
 ///
 /// You can declare Success and Failure Types as Generic Type..
-class SimpleResult<Success, Failure> extends Equatable{
+class SimpleResult<Success, Failure> extends Equatable {
   final bool _isSuccess;
-  final Success _value;
-  final Failure _failure;
+  final Success? _value;
+  final Failure? _failure;
 
-  SimpleResult._({bool isSuccess, Success value, Failure failure})
+  const SimpleResult._(
+      {required bool isSuccess, Success? value, Failure? failure})
       : _isSuccess = isSuccess,
         _value = value,
         _failure = failure;
@@ -33,29 +33,29 @@ class SimpleResult<Success, Failure> extends Equatable{
   bool get isFailure => _isSuccess == false;
 
   /// returns the success value or null if there is a failure.
-  Success get success => _value;
+  Success? get success => _value;
 
   /// returns the failure or null if the state is success.
-  Failure get failure => _failure;
+  Failure? get failure => _failure;
 
   /// Somehow like a switch case statement. You can return a Value from the
   /// functions success and failure.
-  ResultType when<ResultType>(
-      {@required ResultType Function(Success) success,
-      @required ResultType Function(Failure) failure}) {
+  ResultType? when<ResultType>(
+      {required ResultType Function(Success?)? success,
+      required ResultType Function(Failure?)? failure}) {
     if (_isSuccess) {
-      if(success == null){
+      if (success == null) {
         return null;
       }
       return success(_value);
     }
-    if (failure == null){
+    if (failure == null) {
       return null;
     }
     return failure(_failure);
   }
 
-  Success getOrElse<ResultType>(Success Function() successFunction) {
+  Success? getOrElse<ResultType>(Success Function() successFunction) {
     if (_isSuccess) {
       return _value;
     }
@@ -64,12 +64,12 @@ class SimpleResult<Success, Failure> extends Equatable{
 
   /// Use map to convert a Result from one value type to another.
   /// Changes only value type, not the Failure.
-  SimpleResult<ResultType, Failure> map<ResultType>(
-      ResultType Function(Success) mapper) {
+  SimpleResult<ResultType, Failure?> map<ResultType>(
+      ResultType Function(Success?) mapper) {
     if (_isSuccess) {
       return SimpleResult<ResultType, Failure>.success(mapper(_value));
     } else {
-      return SimpleResult<ResultType, Failure>.failure(_failure);
+      return SimpleResult<ResultType, Failure?>.failure(_failure);
     }
   }
 
@@ -78,23 +78,27 @@ class SimpleResult<Success, Failure> extends Equatable{
       Function1<ResultType, Failure> f) {
     if (_isSuccess) {}
   }*/
-  SimpleResult<ResultType, Failure> flatMap<ResultType>(
-      SimpleResult<ResultType, Failure> Function(Success) mapper) {
+  SimpleResult<ResultType, Failure?> flatMap<ResultType>(
+      SimpleResult<ResultType, Failure> Function(Success?) mapper) {
     if (_isSuccess) {
       final result = mapper(_value);
       return result;
     }
-    return SimpleResult<ResultType, Failure>.failure(_failure);
+    return SimpleResult<ResultType, Failure?>.failure(_failure);
   }
 
-  @override  
-  List<Object> get props => [_value, _failure, _isSuccess];
+  @override
+  List<Object?> get props => [_value, _failure, _isSuccess];
 
   @override
   String toString() {
-    if (_isSuccess){
-      return  _value == null ? "Success: (null)" : 'Success: ${_value.toString()}';
+    if (_isSuccess) {
+      return _value == null
+          ? "Success: (null)"
+          : 'Success: ${_value.toString()}';
     }
-    return _failure == null ? 'Failure: (null)' : 'Failure: ${_failure.toString()}';
+    return _failure == null
+        ? 'Failure: (null)'
+        : 'Failure: ${_failure.toString()}';
   }
 }
